// branding-calculator-frontend/src/app/models/constructor-state.model.ts
export interface ConstructorState {
    fullName: string;
    position: string;
    phone: string;
    mobilePhone: string;
    email: string;
    address: string;
    qrCode1: string;
    qrCode1Label: string;
    qrCode2: string;
    qrCode2Label: string;
    qrCodesGenerated: boolean;
    qrCode2Enabled: boolean;
}